Added pufferFish level and some refactoring.
Added an algorithm mostly like a breadth first search to generate the maze for the maze level.  The algorithm is different from a breadth first search because it uses a list instead of a queue and adds the new element at a random position in the list.  This makes for better variety in the maze paths.

When the maze is created, the caller can specify whether to use a bfs or dfs algorithm.

Did minor refactoring of code.

Changed hole in maze level from using a cylindar model to using a flat quad.