Created cat level.
Switched to using the width and height of the window to calculate dimensions of the space to operate in for all mazes.

Created the open area maze.  In this maze, the walls are thinner and the ball can move freely in the cell.

Added maze where one collects things to finish the level.  The things collected (kittens in this case), follow the user's ball (cat in this case) after they are collected.  After all objects are collected, the user can complete the level by going to the hole.