Calculate the maxX and maxY from the VP matrix.
Calculate maxX and maxY from the view and projection matrices.  Use maxX and maxY to generate the mazes so that they take up the entire screen.

make the center of the growing quad level finisher to be the same screen coordinates as the hole.

changed parameters

Have the face of the ball pointing upwards.

Throw up an error dialog if an error occurs instead of putting the error in a text box.