Game should save its state when shutting down.
The game needs to save its state when shutting down.  Needed to restructure things a bit to get this to work.  Now there is a worker thread and a GUI thread.  The worker thread will shut down only when SurfaceHolder.Callback.surfaceDestroyed() is called.  Before shutting down, the worker thread saves the state information to a cbor file.  For now, this information only includes the level the maze was on.  More information (such as what state the level was in) will require more restructuring.

free JNI local references.

Free all JNI local references that are allocated in getTextImage because these references stay around for a while.