refactoring the way save games are done.
Changing data structure used for save game.  The data structures are now called GameSaveData and LevelSaveData.  Each level has its own save data type inherited from LevelSaveData.  GameSaveData just contains version of the save data, the level name, and the screen size.  In JSON, it contains the LevelSaveData (which is level specific).  In C++, they are separate structures.

LevelTracker passes in a LevelSaveData to create a set of level functions (LevelGroup) that will return a level starter, level, and level finisher.  The mazeGraphics can then call the functions when they want to create the starter, level, and finisher.  Currently, only the first level: beginning accepts a LevelSaveData and saves its LevelSaveData when signaled to shutdown.  The rest of the levels just save the GameSaveData (with the idea that they will save their level data with a further checkin).

upgrade gradle

Fixed remainder of copywrites to the current year.