fixed projection matrix in OpenGL
Moved #defines of GLM_FORCE_DEPTH_ZERO_TO_ONE and GLM_FORCE_RADIANS to CMakeLists.txt.  The GLM library uses these to force the depth to zero to one (to be compatible with Vulkan), and to force angles to be measured in radians.

Made getPerspectiveMatrix and getOrthoMatrix functions which take as a parameter what the depth value range should be and whether the y axis should be inverted.

Removed the GLM #defines from the header files.