fixed bounce calculations, mostly
Still needs optimization.  There are still a few cases when it gets a positive dot product of the normal with the velocity.  This value should always be negative since the ball is traveling towards the surface it is bouncing off of.  Also, there are a few cases where it gets that the normal is point in the z direction only.  This should no longer be the case.  In this case, it just bounces back in the direction it came.