use non-buggy surfaces and check for buggy Vulkan
Check to see if Vulkan is buggy.  If so, don't use it and switch to OpenGL.  Also check to see if a 32 bit integer surface is supported in OpenGL without bugs.  If so, use it, else fall back to 255 bit precision surface.

Also fixed a bug where the ball would gitter in the z direction

Also fixed a bug where the ball could get on extruded surfaces if moving slowly.