bug fix: getting stuck at edges.
The ball was getting stuck at the edges.  Only use Level's checkBallBorders to see if the ball is out of bounds.  If it is out of bounds, don't put it right at the edge, put it in a little bit from the edge to make sure it just doesn't hit the same condition again.