New maze level: Movable passage
level with placement of and rotation of cooridors to get the goal.

Users tap cooridor elements to rotate them and drag to move them.  They need to move them into place so that the ball has a path to the goal.

The color of the corridor changes when they are not movable (i.e. the ball has them in its path or they are at the start of the tunnel).

also fixed a bug in model2cbor.cpp that was causing the model to not appear if there were not texture coordinates.

Pass in the projection and view matrices as parameters to the level and level starter create functions. This way we have the option of passing the width and height at other depths to the level.

Cube model generated programatically.

fixed ignore file: no *.o files in the modelobj2cbor directory should be checked in.  Also, the model2cbor and cbor2json files should not be checked in.