New level: rotatable passage.
Uses generated maze to generate a maze and fills each cell with a junction component (turn, straight, t-junction, cross-junction, or dead end).  Then it rotates all the components to a random direction (0, pi/2, pi, or 3pi/2).  The user has to rotate the passages to make a passage to the target.

Moving the components is not allowed (only rotating).

The component placements can't be rotated after the ball goes through them.  If the ball rolls back or rolls around in a loop, the tunnel will be unlocked.  The locked tunnel is indicated as locked by having the texture change.

Separated the logic out of movablePassage.* to new files: movablePassageAlgorithms.*

Changed component models (used for the mouse and gopher level) to indicate edges better.  They are now raised in the middle of the wall to better indicate that a wall is present but keep it separate from the neighboring component.

Moved the algorithms for generating the maze out of maze.* and into generatedMazeAlgorithms.* The algorithm for generating a maze can now be used for other levels that don't inherit from Maze.

Fixed a bug in the finisher for Vulkan (it was using the wrong projection matrix when calculating the width and height of the visible area on the phone at the finisher height.