Redo level loading.
We now use JSON config files to store level information and the sequence of levels.  There is one JSON file for keeping track of the level sequence.  It lists the level name, the description, and the CBOR file used to load the level.  Once the level CBOR file is retrieved from the level sequence file (levels.json), it is used by the level tracker to load the level. Note: the reason levels.json is a JSON file and not a CBOR file is that Java also needs to read this file to display the list of levels.

The level tracker and each levels/*/serializer.cpp are the only places that know how to decode the level CBOR files.  The sequence of events for loading a level are as follows:

(1) The C++ thread starts up and each level type registers itself with the level tracker using a level type name.  The level type name is different from the level name.  The level type indicates the type of level it is like: "generatedMaze".  The level type may be used for multiple levels.  The level name is a unique name describing the level.

(2) The level tracker starts up and checks for a save file.  The save file contains the level name and the save data.  The level name is used to look up the config file in levels.json.  If there is no save data file, then the first level is loaded.

(3) The level tracker decodes the CBOR level config and the CBOR save data.  Then it hands those off to the registered function for that level that takes the json object and finishes decoding it into LevelConfig and LevelSaveData structures.  This function then returns another function with these data in it (using a lambda function).  The level tracker then gives this function, along with similar functions for the level starter and level finisher, to the level sequencer.

(4) The level sequencer calls the functions to generate the level starter, level, and level finisher.

Also did the following:

Save rotatable passage game data.  Save the model and texture indices so that the game board doesn't visably shift that much when reloading from save.

Added saving of movablePassage level information in the save data file.

switched to passing proj, view instead of storing the projection and view matrix in the level tracker.

Reworked directory structure and class names.  Now each level type has a directory under levels.  All the classes have standard names: Level (for the level type), LevelSaveData (for the leve save data), and LevelConfigData (for the level config data).