Starting rewrite of DrawObject/DrawObjectData
By the end of this rewrite, we should be able to:

(1) switch a model matrix to a new texture without deleting the descriptor set or uniform buffers (the descriptor set will have to be updated, of course).
(2) more modular design to make way for having different pipelines for different objects.
(3) uses a model map which will be similar to a texture map, so that models will only need to be loaded once even if they are using different textures.