Support Vulkan for systems with 32 bit CPUs
On 32 bit, non-dispatchable handles are uint64_t and for 64 bit, they are a pointer (in Vulkan).  So, because I wanted to manage the non-dispatchable handles with a shared pointer and on the 32 bit system, the handle would not fit in the space of a pointer, I made a struct to store the handle and allocated it with new and stored that in the shared pointer.  There is an "access the Vulkan handle" function called getVkType<>().