Added objectNoShadows render details.
Added a pipeline to have an object that does not cast shadows.  It will be used for the ball and hole in the dark level.  But it can also be used to not use shadows on slower phones for performance reasons.  This pass it is hard coded that all mazes use no shadows.  Next pass, I'll add an option to enable it from the settings menu.