Add Dark Render Details
Dark Render Details is a render details that attempts to use shadow maps to see if an object would be in the light from a light source or if it would be shaded by another object.  Also, the light source falls off like 1/r^2.  The problem is that it requires a lot of compute power because of the multiple shadow maps required.  It is assumed that the light source is in the maze (and is infact the ball).