Added the ability to read glb files.
Reads Color, Texture Coordinates, Positions, and Vertex Normals.  Face
normals are computed from the positions of each set of three vertices
that make up a triangle and the vertex normals. Modified one model
(end.bend) to have vertex colors selected in blender and had blender
output to a glb file for this model.