Added ability to model2cbor to read GLB files.
Blender will store the vertex colors in a GLB file, however it will
not do so in a OBJ file.  There are a few models where we would
like to have the vertex colors.  Also, GLB format does not have
the face normals and OBJ format does not have the vertex normals.
The vertex and face normals that are missing are now computed in
model2cbor and both are always exported.